+++ src/dryad/format/endf/tabulatedCrossSection.hpp
@@ -7,7 +7,7 @@
 // other includes
 #include "tools/Log.hpp"
 #include "dryad/format/endf/createBoundaries.hpp"
-#include "dryad/format/endf/interpolants.hpp"
+#include "dryad/format/endf/createInterpolants.hpp"
 #include "dryad/TabulatedCrossSection.hpp"
 #include "ENDFtk/section/3.hpp"
 #include "ENDFtk/tree/Section.hpp"
