+++ src/dryad/format/endf/test/CMakeLists.txt
@@ -1,6 +1,7 @@
-add_cpp_test( format.endf.createBoundary   createBoundary.test.cpp )
-add_cpp_test( format.endf.createBoundaries createBoundaries.test.cpp )
-add_cpp_test( format.endf.interpolants     interpolants.test.cpp )
+add_cpp_test( format.endf.createBoundary     createBoundary.test.cpp )
+add_cpp_test( format.endf.createBoundaries   createBoundaries.test.cpp )
+add_cpp_test( format.endf.createInterpolant  createInterpolant.test.cpp )
+add_cpp_test( format.endf.createInterpolants createInterpolants.test.cpp )
 
 add_cpp_test( format.endf.tabulatedCrossSection tabulatedCrossSection.test.cpp )
 add_cpp_test( format.endf.reaction              reaction.test.cpp )
